feat: add checks for existing installations in floorp, vscode, and zsh scripts

+++ install/floorp.sh
@@ -1,4 +1,9 @@
 install() {
+    if dpkg -l | grep -q floorp; then
+        echo "Floorp is already installed."
+        exit
+    fi
+
     curl -fsSL https://ppa.ablaze.one/KEY.gpg | sudo gpg --dearmor -o /usr/share/keyrings/Floorp.gpg
     sudo curl -sS --compressed -o /etc/apt/sources.list.d/Floorp.list 'https://ppa.ablaze.one/Floorp.list'
